shinenigans

+++ utility functions/miscellanous/round-down.js
@@ -1,6 +1,6 @@
 const roundDown = (number) => {
   if (typeof number === "string") {
-    return Math.floor(parseInt(number) * 100) / 100;
+    return Math.floor(parseFloat(number) * 100) / 100;
   } else {
     return Math.floor(number * 100) / 100;
   }
